feat(bulkDelete): refactor delete blocker modal and deletion verification logic
